CRUD機能実装

+++ docker/mysql/sql_init/01_ddl.sql
@@ -8,7 +8,7 @@ CREATE TABLE IF NOT EXISTS sample_schema.todo(
     assagin_person     varchar(20) NOT NULL,
     title              varchar(128) NOT NULL,
     todo_description   varchar(255) NOT NULL,
-    created_at         int NOT NULL,
-    updated_at         int NOT NULL,
+    created_at         bigint NOT NULL,
+    updated_at         bigint NOT NULL,
     INDEX(id)
 );
